Reorganise basic file structure
