+++ public/index.html
@@ -6,7 +6,7 @@
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>ROctopus</title>
+    <title>rocto</title>
     <meta name="description" content="">
     <meta name="author" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -29,7 +29,7 @@
     <section class="intro">
   		<div class="center-content">
   			<div class="intro-content text-center">	
-  				<h2 class="maintitle">ROctopus</h2>
+  				<h2 class="maintitle">rocto</h2>
   				<h1 class="subtitle">coming soon</h1>
           <p id="maincontent"></p>
   			</div>
